fix(Select): fix v-model binding with inline search
