+++ .changeset/pink-jars-strive.md
@@ -1,0 +1,5 @@
+---
+"@wisemen/vue-core-components": patch
+---
+
+Select: Resolved an issue where v-model did not work correctly when using inline search.
